+++ src/Symfony/Component/VarDumper/Caster/ReflectionCaster.php
@@ -98,6 +98,7 @@ class ReflectionCaster
 
         self::addMap($a, $c, array(
             'returnsReference' => 'returnsReference',
+            'returnType' => 'getReturnType',
             'class' => 'getClosureScopeClass',
             'this' => 'getClosureThis',
         ));